+++ examples/block.spls-examples.R
@@ -11,7 +11,7 @@ design
 # set number of component per data set
 ncomp = c(2)
 # set number of variables to select, per component and per data set (this is set arbitrarily)
-list.keepX = list(mrna = rep(20, 2), mirna = rep(10,2))
+list.keepX = list(mrna = rep(5, 2), mirna = rep(5,2))
 list.keepY = c(rep(10, 2))
 
 TCGA.block.spls = block.spls(X = data, Y = breast.TCGA$data.train$protein,
@@ -22,6 +22,16 @@ plotIndiv(TCGA.block.spls, group =  breast.TCGA$data.train$subtype, ind.names = 
 # illustrates coefficient weights in each block
 plotLoadings(TCGA.block.spls, ncomp = 1)
 plotVar(TCGA.block.spls, style = 'graphics', legend = TRUE)
+
+## plot markers (selected markers) for mrna and mirna
+group <- breast.TCGA$data.train$subtype
+# mrna: show each selected feature separately and group by subtype
+plotMarkers(object = TCGA.block.spls, comp = 1, block = 'mrna', group = group)
+# mrna: aggregate all selected features, separate by loadings signs and group by subtype
+plotMarkers(object = TCGA.block.spls, comp = 1, block = 'mrna', group = group, global = TRUE)
+# proteins
+plotMarkers(object = TCGA.block.spls, comp = 1, block = 'Y', group = group)
+
 \dontrun{
 network(TCGA.block.spls)
 }